Refactor: Move patch_fx_custom_device to musa backend for better organization
